Refactor modules: Remove hardcoded magical array size. [1/4]
AirlineData
AnimalData
BookData
ColorData
CommerceData
CompanyData
ComputerData
