--- src/modules/color/ColorData.cpp
@@ -1,14 +1,0 @@
-#include "ColorData.h"
-
-#include <array>
-#include <string_view>
-
-namespace faker::color
-{
-const std::array<std::string_view, 31> colors = {
-    "red",      "green",  "blue",    "yellow", "purple", "mint green", "teal",      "white",
-    "black",    "orange", "pink",    "grey",   "maroon", "violet",     "turquoise", "tan",
-    "sky blue", "salmon", "plum",    "orchid", "olive",  "magenta",    "lime",      "ivory",
-    "indigo",   "gold",   "fuchsia", "cyan",   "azure",  "lavender",   "silver",
-};
-}
