+++ src/modules/company/CompanyData.h
@@ -5,13 +5,603 @@
 
 namespace faker::company
 {
-extern const std::array<std::string_view, 65> buzzAdjectives;
-extern const std::array<std::string_view, 40> buzzNouns;
-extern const std::array<std::string_view, 60> buzzVerbs;
-extern const std::array<std::string_view, 100> catchPhraseAdjectives;
-extern const std::array<std::string_view, 101> catchPhraseDescriptors;
-extern const std::array<std::string_view, 104> catchPhraseNouns;
-extern const std::array<std::string_view, 8> companyTypes;
-extern const std::array<std::string_view, 147> companyIndustries;
-extern const std::array<std::string_view, 3> companySuffixes;
+const auto buzzAdjectives = std::to_array<std::string_view>({
+    "clicks-and-mortar",
+    "value-added",
+    "vertical",
+    "proactive",
+    "robust",
+    "revolutionary",
+    "scalable",
+    "leading-edge",
+    "innovative",
+    "intuitive",
+    "strategic",
+    "e-business",
+    "mission-critical",
+    "sticky",
+    "one-to-one",
+    "24/7",
+    "end-to-end",
+    "global",
+    "B2B",
+    "B2C",
+    "granular",
+    "frictionless",
+    "virtual",
+    "viral",
+    "dynamic",
+    "24/365",
+    "best-of-breed",
+    "killer",
+    "magnetic",
+    "bleeding-edge",
+    "web-enabled",
+    "interactive",
+    "dot-com",
+    "sexy",
+    "back-end",
+    "real-time",
+    "efficient",
+    "front-end",
+    "distributed",
+    "seamless",
+    "extensible",
+    "turn-key",
+    "world-class",
+    "open-source",
+    "cross-platform",
+    "cross-media",
+    "synergistic",
+    "bricks-and-clicks",
+    "out-of-the-box",
+    "enterprise",
+    "integrated",
+    "impactful",
+    "wireless",
+    "transparent",
+    "next-generation",
+    "cutting-edge",
+    "user-centric",
+    "visionary",
+    "customized",
+    "ubiquitous",
+    "plug-and-play",
+    "collaborative",
+    "compelling",
+    "holistic",
+    "rich",
+});
+
+const auto buzzNouns = std::to_array<std::string_view>({
+    "synergies",
+    "paradigms",
+    "markets",
+    "partnerships",
+    "infrastructures",
+    "platforms",
+    "initiatives",
+    "channels",
+    "eyeballs",
+    "communities",
+    "ROI",
+    "solutions",
+    "action-items",
+    "portals",
+    "niches",
+    "technologies",
+    "content",
+    "supply-chains",
+    "convergence",
+    "relationships",
+    "architectures",
+    "interfaces",
+    "e-markets",
+    "e-commerce",
+    "systems",
+    "bandwidth",
+    "models",
+    "mindshare",
+    "deliverables",
+    "users",
+    "schemas",
+    "networks",
+    "applications",
+    "metrics",
+    "e-business",
+    "functionalities",
+    "experiences",
+    "web services",
+    "methodologies",
+    "blockchains"
+    "lifetime value",
+});
+
+const auto buzzVerbs = std::to_array<std::string_view>({
+    "implement",     "utilize",     "integrate",       "streamline", "optimize",       "evolve",     "transform",
+    "embrace",       "enable",      "orchestrate",     "leverage",   "reinvent",       "aggregate",  "architect",
+    "enhance",       "incentivize", "morph",           "empower",    "envisioneer",    "monetize",   "harness",
+    "facilitate",    "seize",       "disintermediate", "synergize",  "strategize",     "deploy",     "brand",
+    "grow",          "target",      "syndicate",       "synthesize", "deliver",        "mesh",       "incubate",
+    "engage",        "maximize",    "benchmark",       "expedite",   "reintermediate", "whiteboard", "visualize",
+    "repurpose",     "innovate",    "scale",           "unleash",    "drive",          "extend",     "engineer",
+    "revolutionize", "generate",    "exploit",         "transition", "e-enable",       "iterate",    "cultivate",
+    "matrix",        "productize",  "redefine",
+});
+
+const auto catchPhraseAdjectives = std::to_array<std::string_view>({
+    "Adaptive",
+    "Advanced",
+    "Ameliorated",
+    "Assimilated",
+    "Automated",
+    "Balanced",
+    "Business-focused",
+    "Centralized",
+    "Cloned",
+    "Compatible",
+    "Configurable",
+    "Cross-group",
+    "Cross-platform",
+    "Customer-focused",
+    "Customizable",
+    "Decentralized",
+    "De-engineered",
+    "Devolved",
+    "Digitized",
+    "Distributed",
+    "Diverse",
+    "Down-sized",
+    "Enhanced",
+    "Enterprise-wide",
+    "Ergonomic",
+    "Exclusive",
+    "Expanded",
+    "Extended",
+    "Face to face",
+    "Focused",
+    "Front-line",
+    "Fully-configurable",
+    "Function-based",
+    "Fundamental",
+    "Future-proofed",
+    "Grass-roots",
+    "Horizontal",
+    "Implemented",
+    "Innovative",
+    "Integrated",
+    "Intuitive",
+    "Inverse",
+    "Managed",
+    "Mandatory",
+    "Monitored",
+    "Multi-channelled",
+    "Multi-lateral",
+    "Multi-layered",
+    "Multi-tiered",
+    "Networked",
+    "Object-based",
+    "Open-architected",
+    "Open-source",
+    "Operative",
+    "Optimized",
+    "Optional",
+    "Organic",
+    "Organized",
+    "Persevering",
+    "Persistent",
+    "Phased",
+    "Polarised",
+    "Pre-emptive",
+    "Proactive",
+    "Profit-focused",
+    "Profound",
+    "Programmable",
+    "Progressive",
+    "Public-key",
+    "Quality-focused",
+    "Reactive",
+    "Realigned",
+    "Re-contextualized",
+    "Re-engineered",
+    "Reduced",
+    "Reverse-engineered",
+    "Right-sized",
+    "Robust",
+    "Seamless",
+    "Secured",
+    "Self-enabling",
+    "Sharable",
+    "Stand-alone",
+    "Streamlined",
+    "Switchable",
+    "Synchronised",
+    "Synergistic",
+    "Synergized",
+    "Team-oriented",
+    "Total",
+    "Triple-buffered",
+    "Universal",
+    "Up-sized",
+    "Upgradable",
+    "User-centric",
+    "User-friendly",
+    "Versatile",
+    "Virtual",
+    "Visionary",
+    "Vision-oriented",
+});
+
+const auto catchPhraseDescriptors = std::to_array<std::string_view>({
+    "24 hour",
+    "24/7",
+    "3rd generation",
+    "4th generation",
+    "5th generation",
+    "6th generation",
+    "actuating",
+    "analyzing",
+    "asymmetric",
+    "asynchronous",
+    "attitude-oriented",
+    "background",
+    "bandwidth-monitored",
+    "bi-directional",
+    "bifurcated",
+    "bottom-line",
+    "clear-thinking",
+    "client-driven",
+    "client-server",
+    "coherent",
+    "cohesive",
+    "composite",
+    "context-sensitive",
+    "contextually-based",
+    "content-based",
+    "dedicated",
+    "demand-driven",
+    "didactic",
+    "directional",
+    "discrete",
+    "disintermediate",
+    "dynamic",
+    "eco-centric",
+    "empowering",
+    "encompassing",
+    "even-keeled",
+    "executive",
+    "explicit",
+    "exuding",
+    "fault-tolerant",
+    "foreground",
+    "fresh-thinking",
+    "full-range",
+    "global",
+    "grid-enabled",
+    "heuristic",
+    "high-level",
+    "holistic",
+    "homogeneous",
+    "human-resource",
+    "hybrid",
+    "impactful",
+    "incremental",
+    "intangible",
+    "interactive",
+    "intermediate",
+    "leading edge",
+    "local",
+    "logistical",
+    "maximized",
+    "methodical",
+    "mission-critical",
+    "mobile",
+    "modular",
+    "motivating",
+    "multimedia",
+    "multi-state",
+    "multi-tasking",
+    "national",
+    "needs-based",
+    "neutral",
+    "next generation",
+    "non-volatile",
+    "object-oriented",
+    "optimal",
+    "optimizing",
+    "radical",
+    "real-time",
+    "reciprocal",
+    "regional",
+    "responsive",
+    "scalable",
+    "secondary",
+    "solution-oriented",
+    "stable",
+    "static",
+    "systematic",
+    "systemic",
+    "system-worthy",
+    "tangible",
+    "tertiary",
+    "transitional",
+    "uniform",
+    "upward-trending",
+    "user-facing",
+    "value-added",
+    "web-enabled",
+    "well-modulated",
+    "zero administration",
+    "zero defect",
+    "zero tolerance",
+});
+
+const auto catchPhraseNouns = std::to_array<std::string_view>({
+    "ability",
+    "access",
+    "adapter",
+    "algorithm",
+    "alliance",
+    "analyzer",
+    "application",
+    "approach",
+    "architecture",
+    "archive",
+    "artificial intelligence",
+    "array",
+    "attitude",
+    "benchmark",
+    "budgetary management",
+    "capability",
+    "capacity",
+    "challenge",
+    "circuit",
+    "collaboration",
+    "complexity",
+    "concept",
+    "conglomeration",
+    "contingency",
+    "core",
+    "customer loyalty",
+    "database",
+    "data-warehouse",
+    "definition",
+    "emulation",
+    "encoding",
+    "encryption",
+    "extranet",
+    "firmware",
+    "flexibility",
+    "focus group",
+    "forecast",
+    "frame",
+    "framework",
+    "function",
+    "functionalities",
+    "Graphic Interface",
+    "groupware",
+    "Graphical User Interface",
+    "hardware",
+    "help-desk",
+    "hierarchy",
+    "hub",
+    "implementation",
+    "info-mediaries",
+    "infrastructure",
+    "initiative",
+    "installation",
+    "instruction set",
+    "interface",
+    "internet solution",
+    "intranet",
+    "knowledge user",
+    "knowledge base",
+    "local area network",
+    "leverage",
+    "matrices",
+    "matrix",
+    "methodology",
+    "middleware",
+    "migration",
+    "model",
+    "moderator",
+    "monitoring",
+    "moratorium",
+    "neural-net",
+    "open architecture",
+    "open system",
+    "orchestration",
+    "paradigm",
+    "parallelism",
+    "policy",
+    "portal",
+    "pricing structure",
+    "process improvement",
+    "product",
+    "productivity",
+    "project",
+    "projection",
+    "protocol",
+    "secured line",
+    "service-desk",
+    "software",
+    "solution",
+    "standardization",
+    "strategy",
+    "structure",
+    "success",
+    "superstructure",
+    "support",
+    "synergy",
+    "system engine",
+    "task-force",
+    "throughput",
+    "time-frame",
+    "toolset",
+    "utilisation",
+    "website",
+    "workforce",
+});
+
+const auto companyTypes = std::to_array<std::string_view>({
+    "Public Company", "Educational Institution", "Self-Employed",  "Government Agency",
+    "Nonprofit",      "Sole Proprietorship",     "Privately Held", "Partnership",
+});
+
+const auto companyIndustries = std::to_array<std::string_view>({
+    "Defense & Space",
+    "Computer Hardware",
+    "Computer Software",
+    "Computer Networking",
+    "Internet",
+    "Semiconductors",
+    "Telecommunications",
+    "Law Practice",
+    "Legal Services",
+    "Management Consulting",
+    "Biotechnology",
+    "Medical Practice",
+    "Hospital & Health Care",
+    "Pharmaceuticals",
+    "Veterinary",
+    "Medical Devices",
+    "Cosmetics",
+    "Apparel & Fashion",
+    "Sporting Goods",
+    "Tobacco",
+    "Supermarkets",
+    "Food Production",
+    "Consumer Electronics",
+    "Consumer Goods",
+    "Furniture",
+    "Retail",
+    "Entertainment",
+    "Gambling & Casinos",
+    "Leisure, Travel & Tourism",
+    "Hospitality",
+    "Restaurants",
+    "Sports",
+    "Food & Beverages",
+    "Motion Pictures and Film",
+    "Broadcast Media",
+    "Museums and Institutions",
+    "Fine Art",
+    "Performing Arts",
+    "Recreational Facilities and Services",
+    "Banking",
+    "Insurance",
+    "Financial Services",
+    "Real Estate",
+    "Investment Banking",
+    "Investment Management",
+    "Accounting",
+    "Construction",
+    "Building Materials",
+    "Architecture & Planning",
+    "Civil Engineering",
+    "Aviation & Aerospace",
+    "Automotive",
+    "Chemicals",
+    "Machinery",
+    "Mining & Metals",
+    "Oil & Energy",
+    "Shipbuilding",
+    "Utilities",
+    "Textiles",
+    "Paper & Forest Products",
+    "Railroad Manufacture",
+    "Farming",
+    "Ranching",
+    "Dairy",
+    "Fishery",
+    "Primary / Secondary Education",
+    "Higher Education",
+    "Education Management",
+    "Research",
+    "Military",
+    "Legislative Office",
+    "Judiciary",
+    "International Affairs",
+    "Government Administration",
+    "Executive Office",
+    "Law Enforcement",
+    "Public Safety",
+    "Public Policy",
+    "Marketing and Advertising",
+    "Newspapers",
+    "Publishing",
+    "Printing",
+    "Information Services",
+    "Libraries",
+    "Environmental Services",
+    "Package / Freight Delivery",
+    "Individual & Family Services",
+    "Religious Institutions",
+    "Civic & Social Organization",
+    "Consumer Services",
+    "Transportationg / Trucking / Railroad",
+    "Warehousing",
+    "Airlines / Aviation",
+    "Maritime",
+    "Information Technology and Services",
+    "Market Research",
+    "Public Relations and Communications",
+    "Design",
+    "Nonprofit Organization Management",
+    "Fund-Raising",
+    "Program Development",
+    "Writing and Editing",
+    "Staffing and Recruiting",
+    "Professional Training & Coaching",
+    "Venture Capital & Private Equity",
+    "Political Organization",
+    "Translation and Localization",
+    "Computer Games",
+    "Events Services",
+    "Arts and Crafts",
+    "Electrical / Electronic Manufacturing",
+    "Online Media",
+    "Nanotechnology",
+    "Music",
+    "Logistics and Supply Chain",
+    "Plastics",
+    "Computer & Network Security",
+    "Wireless",
+    "Alternative Dispute Resolution",
+    "Security and Investigations",
+    "Facilities Services",
+    "Outsourcing / Offshoring",
+    "Health, Wellness and Fitness",
+    "Alternative Medicine",
+    "Media Production",
+    "Animation",
+    "Commercial Real Estate",
+    "Capital Markets",
+    "Think Tanks",
+    "Philanthropy",
+    "E-Learning",
+    "Wholesale",
+    "Import and Export",
+    "Mechanical or Industrial Engineering",
+    "Photography",
+    "Human Resources",
+    "Business Supplies and Equipment",
+    "Mental Health Care",
+    "Graphic Design",
+    "International Trade and Development",
+    "Wine and Spirits",
+    "Luxury Goods & Jewelry",
+    "Renewables & Environment",
+    "Glass, Ceramics & Concrete",
+    "Packaging and Containers",
+    "Industrial Automation",
+    "Government Relations",
+});
+
+const auto companySuffixes = std::to_array<std::string_view>({
+    "Inc",
+    "LLC",
+    "Group",
+});
+
 }
